students/k3344/Solovyova_Polina/lr1/book_exchange/requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-dotenv
pydantic-settings